main game part was made

+++ core/src/com/samsung/game/items/weapon/Weapon.java
@@ -4,9 +4,9 @@ import com.samsung.game.items.Item;
 import com.samsung.game.engine.Damage;
 
 public abstract class Weapon extends Item implements Damage {
-    private float damage_min, damage_max;
+    private int damage_min, damage_max;
 
-    public Weapon(float damage_min, float damage_max) {
+    public Weapon(int damage_min, int damage_max) {
         this.damage_max = damage_max;
         this.damage_min = damage_min;
     }
